Fix brush crash and scrollRequest float type error
- canvas.py: consolidate QPainter into single instance per _paintBrushLine
  (was creating/destroying painter per interpolated point -> crash under load)
- canvas.py: fix prevPoint None guard in brush mouseMoveEvent
- canvas.py: replace _paintBrushAt with unified _paintBrushLine(p, p) for click
- canvas.py: extract _maskToNumpy() helper, remove byteCount() usage
- app.py: fix scrollRequest float->int cast (Alt+scroll crash)

+++ labelme/app.py
@@ -1071,7 +1071,7 @@ class MainWindow(QtWidgets.QMainWindow, WindowMixin):
     def scrollRequest(self, delta, orientation):
         units = - delta * 0.1  # natural scroll
         bar = self.scrollBars[orientation]
-        bar.setValue(bar.value() + bar.singleStep() * units)
+        bar.setValue(int(bar.value() + bar.singleStep() * units))
 
     def setZoom(self, value):
         self.actions.fitWidth.setChecked(False)
